+++ package.json
@@ -13,22 +13,22 @@
         "format:check": "prettier --check ."
     },
     "dependencies": {
-        "@astrojs/markdown-remark": "^7.2.1",
+        "@astrojs/markdown-satteri": "^0.3.3",
         "@astrojs/starlight": "^0.41.3",
+        "@system76/satteri-relative-markdown-links": "^1.0.0",
         "@types/mdast": "^4.0.4",
-        "astro": "^7.0.3",
-        "astro-rehype-relative-markdown-links": "^0.19.0",
+        "astro": "^7.0.6",
         "remark": "^15.0.1",
+        "satteri": "^0.9.4",
         "sharp": "^0.35.3",
-        "unified": "^11.0.5",
         "vite-plugin-static-copy": "^4.1.1"
     },
     "devDependencies": {
         "@types/js-yaml": "^4.0.9",
-        "js-yaml": "^5.2.1",
+        "js-yaml": "^4.3.0",
         "prettier": "^3.9.4",
         "prettier-plugin-astro": "^0.14.1",
         "prettier-plugin-organize-imports": "^4.3.0"
     },
-    "packageManager": "pnpm@11.9.0+sha512.bd682d5d03fe525ef7c9fd6780c6884d1e756ac4c9c9fe00c538782824310dcf90e3ddc4f53835f06dfaebd5085e41855e0bcbb3b60de2ac5bbab89e5036f03b"
+    "packageManager": "pnpm@11.10.0+sha512.0b7f8b98060031904c017e3a41eb187a16d40eeb829b95c4f8cb03681761fc4ab53dd219115b9b447f4dce1a05a214764461e7d3703392a9f32f9511ce8c86c8"
 }
